Move `RelationKind` to AST.
