+++ include/Index/SymbolIndex.h
@@ -3,8 +3,8 @@
 #include "Shared.h"
 #include "ArrayView.h"
 #include "AST/SymbolKind.h"
+#include "AST/RelationKind.h"
 #include "Basic/SourceCode.h"
-#include "Basic/RelationKind.h"
 #include "Support/JSON.h"
 
 namespace clice {
