+++ src/Feature/SemanticTokens.cpp
@@ -1,6 +1,7 @@
 #include "AST/Semantic.h"
 #include "Index/Shared.h"
 #include "Feature/SemanticTokens.h"
+#include "Support/Compare.h"
 
 namespace clice::feature {
