+++ src/Index/SymbolIndex.cpp
@@ -6,7 +6,7 @@
 #include "Basic/SourceCode.h"
 #include "Index/SymbolIndex.h"
 #include "Support/Binary.h"
-
+#include "Support/Compare.h"
 #include "clang/Index/USRGeneration.h"
 
 namespace clice::index {
